Main menu layout

+++ app/src/main/java/comp3350/studentlifesimulator/presentation/MainActivity.java
@@ -2,11 +2,16 @@ package comp3350.studentlifesimulator.presentation;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.app.AlertDialog;
+
 import android.content.Context;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.res.AssetManager;
 import android.os.Bundle;
+import android.view.View;
 import android.widget.Button;
+
 
 import com.example.studentlifesimulator.R;
 
@@ -18,7 +23,9 @@ import java.io.InputStreamReader;
 import comp3350.studentlifesimulator.application.Main;
 
 public class MainActivity extends AppCompatActivity {
-    Button newGame;
+    Button newGame, continueGame, credits, howToPlay;
+    boolean newGamePressed = false;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -29,11 +36,57 @@ public class MainActivity extends AppCompatActivity {
         Main.openDBAccess();
 
         newGame = findViewById(R.id.newGameButton);
-        newGame.setOnClickListener(view -> switchActivity());
+        continueGame = findViewById(R.id.continueButton);
+        credits = findViewById(R.id.creditsButton);
+        howToPlay = findViewById(R.id.howToPlayButton);
+
+        if (Main.checkPreviousData()){
+            continueGame.setVisibility(View.VISIBLE);
+        }
+
+        newGame.setOnClickListener(v -> {
+            newGamePressed = true;
+            copyDatabaseToDevice();
+            switchActivity();
+
+        });//
+        continueGame.setOnClickListener(View -> switchActivity());
+
+        credits.setOnClickListener(v -> {
+            AlertDialog alertDialog = new AlertDialog.Builder(MainActivity.this).create(); //Read Update
+            alertDialog.setTitle("Credits");
+            alertDialog.setMessage(getString(R.string.dialog_credits));
+
+            alertDialog.setButton("Close", new DialogInterface.OnClickListener() {
+                public void onClick(DialogInterface dialog, int whichButton) {
+
+                }
+            });
+            alertDialog.show();
+        });
+
+        howToPlay.setOnClickListener(v ->{
+            AlertDialog alertDialog = new AlertDialog.Builder(MainActivity.this).create();
+            alertDialog.setTitle("How To Play");
+            alertDialog.setMessage(getString(R.string.how_to_play));
+
+            //noinspection deprecation
+            alertDialog.setButton("I Know", new DialogInterface.OnClickListener() {
+                public void onClick(DialogInterface dialog, int whichButton) {
+
+                }
+            });
+            alertDialog.show();
+
+        });
+
     }
 
-    private void switchActivity() {
+    private void switchActivity() {//nextActivity = new Intent(this, ApartmentActivity.class);
         Intent nextActivity;
+
+        System.out.println("Data " + Main.checkPreviousData());
+
 
         if (Main.checkPreviousData()) {
             nextActivity = new Intent(this, ApartmentActivity.class);
@@ -41,7 +94,6 @@ public class MainActivity extends AppCompatActivity {
         else {
             nextActivity = new Intent(this, CoursesActivity.class);
         }
-
         startActivity(nextActivity);
     }
 
@@ -78,13 +130,22 @@ public class MainActivity extends AppCompatActivity {
         FileWriter out;
         int count;
 
+
         for (String asset : assets) {
             components = asset.split("/");
             copyPath = targetDirectory.toString() + "/" + components[components.length - 1];
             buffer = new char[1024];
             outFile = new File(copyPath);
 
-            if (!outFile.exists()) {
+            System.out.println(outFile.exists());
+            System.out.println(newGamePressed);
+
+            if (outFile.exists() && newGamePressed){
+                outFile.delete();
+                System.out.println("In if");
+            }
+
+            if (!outFile.exists() && !newGamePressed) {
                 in = new InputStreamReader(assetManager.open(asset));
                 out = new FileWriter(outFile);
 
@@ -99,4 +160,5 @@ public class MainActivity extends AppCompatActivity {
             }
         }
     }
+
 }
